+++ src/OmniSharp.Roslyn/RoslynInternals/Analyzers/RoslynAnalyzerServices.cs
@@ -1,8 +1,8 @@
 #nullable enable
 
 using System;
+using System.IO;
 using System.Linq;
-using System.Reflection;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
 using OmniSharp.Roslyn.Reflection;
@@ -14,22 +14,32 @@ namespace OmniSharp.Roslyn.RoslynInternals.Analyzers
         public static IAnalyzerAssemblyLoader CreateShadowCopyAnalyzerAssemblyLoader()
         {
             var type = RoslynReflection.GetType(RoslynReflection.CodeAnalysisAssembly, "Microsoft.CodeAnalysis.AnalyzerAssemblyLoader");
-            if (Type.GetType("Mono.Runtime") is not null)
+            if (!OperatingSystem.IsWindows())
             {
                 return (IAnalyzerAssemblyLoader)Activator.CreateInstance(type, nonPublic: true)!;
             }
 
-            var clean = type.GetMethods(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static)
-                .SingleOrDefault(m => m.Name == "CleanLegacyShadowCopyDirectoryIfNeeded" &&
-                                      m.GetParameters().Length == 1);
+            var clean = RoslynReflection.GetMethod(
+                type,
+                "CleanLegacyShadowCopyDirectoryIfNeeded",
+                parameterCount: 1,
+                isStatic: true);
             var create = RoslynReflection.GetMethod(type, "CreateNonLockingLoader",
-                m => m.IsStatic && m.GetParameters().Length >= 1 &&
+                m => m.IsStatic &&
+                     m.GetParameters().FirstOrDefault()?.ParameterType == typeof(string) &&
                      m.GetParameters().Skip(1).All(p => p.IsOptional));
-            var root = System.IO.Path.Combine(System.IO.Path.GetTempPath(), "CodeAnalysis");
-            if (clean is not null)
-                RoslynReflection.Invoke(clean, null, System.IO.Path.Combine(root, "OmnisharpAnalyzerShadowCopies"));
-            var arguments = Enumerable.Repeat<object>(Type.Missing, create.GetParameters().Length).ToArray();
-            arguments[0] = System.IO.Path.Combine(root, "OmnisharpAnalyzerPathResolver");
+
+            var root = Path.Combine(Path.GetTempPath(), "CodeAnalysis");
+            RoslynReflection.Invoke(clean, null, Path.Combine(root, "OmnisharpAnalyzerShadowCopies"));
+
+            var parameters = create.GetParameters();
+            var arguments = parameters
+                .Select(parameter => parameter.ParameterType.IsValueType
+                    ? Activator.CreateInstance(parameter.ParameterType)
+                    : null)
+                .ToArray();
+            arguments[0] = Path.Combine(root, "OmnisharpAnalyzerPathResolver");
+
             return RoslynReflection.Invoke<IAnalyzerAssemblyLoader>(create, null, arguments);
         }
     }