+++ tests/OmniSharp.DotNetTest.Tests/DiscoverTestFacts.cs
@@ -21,7 +21,7 @@ namespace OmniSharp.DotNetTest.Tests
             await DiscoverTestsAsync(
                 XunitTestProject,
                 testFramework: "xunit",
-                targetFrameworkVersion: ".NETCoreApp,Version=v6.0",
+                targetFrameworkVersion: ".NETCoreApp,Version=v8.0",
                 new Test
                 {
                     FullyQualifiedName = "Main.Test.MainTest.UsesDisplayName",
@@ -36,7 +36,7 @@ namespace OmniSharp.DotNetTest.Tests
             await DiscoverTestsAsync(
                 NUnitTestProject,
                 testFramework: "nunit",
-                targetFrameworkVersion: ".NETCoreApp,Version=v6.0",
+                targetFrameworkVersion: ".NETCoreApp,Version=v8.0",
                 new Test
                 {
                     FullyQualifiedName = "Main.Test.MainTest.Test",
@@ -51,7 +51,7 @@ namespace OmniSharp.DotNetTest.Tests
             await DiscoverTestsAsync(
                 MSTestProject,
                 testFramework: "mstest",
-                targetFrameworkVersion: ".NETCoreApp,Version=v6.0",
+                targetFrameworkVersion: ".NETCoreApp,Version=v8.0",
                 new Test
                 {
                     FullyQualifiedName = "Main.Test.MainTest.CheckStandardOutput",