+++ tests/OmniSharp.MSBuild.Tests/ProjectWithComplexAnalyzersTests.cs
@@ -15,8 +15,7 @@ namespace OmniSharp.MSBuild.Tests
         {
         }
 
-        // possible thread starvation on *nix when running in Azure DevOps
-        [ConditionalFact(typeof(WindowsOnly))]
+        [Fact]
         public async Task CanLoadComplexAnalyzers()
         {
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("ProjectWithComplexAnalyzers"))