+++ tests/OmniSharp.Roslyn.CSharp.Tests/CompletionFacts.cs
@@ -633,13 +633,8 @@ class FooChild : Foo
 
             using var host = GetAsyncCompletionAndImportCompletionHost();
             var completions = await FindCompletionsAsync(filename, source, host);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "Test(string text)", "Test(string text, string moreText)", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "Test(string text)", "Test(string text, string moreText)", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
             Assert.Equal(new[] { "Equals", "GetHashCode", "Test", "Test", "ToString" },
                          completions.Items.Select(c => c.TextEdit.NewText));
 
@@ -676,13 +671,8 @@ class FooChild : Foo
 ";
 
             var completions = await FindCompletionsAsync(filename, source, SharedOmniSharpTestHost);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "Test(string text)", "Test(string text, string moreText)", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "Test(string text)", "Test(string text, string moreText)", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "public override bool Equals(object obj)\n    {\n        return base.Equals(obj);$0\n    \\}",
                                  "public override int GetHashCode()\n    {\n        return base.GetHashCode();$0\n    \\}",
@@ -740,13 +730,8 @@ namespace N3
 
             using var host = GetAsyncCompletionAndImportCompletionHost();
             var completions = await FindCompletionsAsync(filename, source, host);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "GetN1()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "GetN1()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
             Assert.Equal(new[] { "Equals", "GetHashCode", "GetN1", "ToString" },
                          completions.Items.Select(c => c.TextEdit.NewText));
 
@@ -790,13 +775,8 @@ namespace N3
 }";
 
             var completions = await FindCompletionsAsync(filename, source, SharedOmniSharpTestHost);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "GetN1()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "GetN1()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "public override bool Equals(object obj)\n        {\n            return base.Equals(obj);$0\n        \\}",
                                  "public override int GetHashCode()\n        {\n            return base.GetHashCode();$0\n        \\}",
@@ -839,13 +819,8 @@ class C
 
             using var host = GetAsyncCompletionAndImportCompletionHost();
             var completions = await FindCompletionsAsync(filename, source, host);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
             Assert.Equal(new[] { "Equals", "GetHashCode", "ToString" },
                          completions.Items.Select(c => c.TextEdit.NewText));
 
@@ -875,13 +850,8 @@ class C
 }";
 
             var completions = await FindCompletionsAsync(filename, source, SharedOmniSharpTestHost);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "bool Equals(object obj)\n    {\n        return base.Equals(obj);$0\n    \\}",
                                  "int GetHashCode()\n    {\n        return base.GetHashCode();$0\n    \\}",
@@ -912,13 +882,8 @@ class C
 
             using var host = GetAsyncCompletionAndImportCompletionHost();
             var completions = await FindCompletionsAsync(filename, source, host);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "Equals", "GetHashCode", "ToString" },
                          completions.Items.Select(c => c.TextEdit.NewText));
@@ -948,13 +913,8 @@ class C
 }";
 
             var completions = await FindCompletionsAsync(filename, source, SharedOmniSharpTestHost);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "public override bool Equals(object obj)\n    {\n        return base.Equals(obj);$0\n    \\}",
                                  "public override int GetHashCode()\n    {\n        return base.GetHashCode();$0\n    \\}",
@@ -978,13 +938,8 @@ class C
 
             using var host = GetAsyncCompletionAndImportCompletionHost();
             var completions = await FindCompletionsAsync(filename, source, host);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "Equals" },
                          completions.Items.Select(c => c.TextEdit.NewText));
@@ -1015,13 +970,8 @@ class C
 }";
 
             var completions = await FindCompletionsAsync(filename, source, SharedOmniSharpTestHost);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "Equals(object obj)\n    {\n        return base.Equals(obj);$0\n    \\}" },
                          completions.Items.Select(c => c.TextEdit.NewText));
@@ -1051,13 +1001,8 @@ class Derived : Base
 
             using var host = GetAsyncCompletionAndImportCompletionHost();
             var completions = await FindCompletionsAsync(filename, source, host);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "Test()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "Test()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "Equals", "GetHashCode", "Test", "ToString" },
                          completions.Items.Select(c => c.TextEdit.NewText));
@@ -1093,13 +1038,8 @@ class Derived : Base
 }";
 
             var completions = await FindCompletionsAsync(filename, source, SharedOmniSharpTestHost);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "Test()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "Test()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "public override bool Equals(object obj)\n    {\n        return base.Equals(obj);$0\n    \\}",
                                  "public override int GetHashCode()\n    {\n        return base.GetHashCode();$0\n    \\}",
@@ -1581,13 +1521,8 @@ class C
 
             using var host = GetAsyncCompletionAndImportCompletionHost();
             var completions = await FindCompletionsAsync(filename, source, host);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "Equals", "GetHashCode", "ToString" },
                          completions.Items.Select(c => c.TextEdit.NewText));
@@ -1618,13 +1553,8 @@ class C
 }";
 
             var completions = await FindCompletionsAsync(filename, source, SharedOmniSharpTestHost);
-#if NETCOREAPP
             Assert.Equal(new[] { "Equals(object? obj)", "GetHashCode()", "ToString()" },
                          completions.Items.Select(c => c.Label));
-#else
-            Assert.Equal(new[] { "Equals(object obj)", "GetHashCode()", "ToString()" },
-                         completions.Items.Select(c => c.Label));
-#endif
 
             Assert.Equal(new[] { "public override bool Equals(object obj)\n    {\n        return base.Equals(obj);$0\n    \\}",
                                  "public override int GetHashCode()\n    {\n        return base.GetHashCode();$0\n    \\}",